Bug #26399073: MYSQL DOES NOT COMPILE WITH CLANG ON WINDOWS [noclose]
Fix a performance regression; InnoDB has a system where the performance
schema key is decided by __FILE__, and the patch that extended the
basename function to work with both / and \ made these allocations
slower (as it uses std::string, which can allocate on the heap).

However, there is no reason whatsoever that this lookup needs to be done
runtime in a map. Change it to use a constexpr function; it goes through
the array linearly, but it's small and it's compile-time. The end result
is that the compiled code stores a PSI key directly, instead of taking
__FILE__, removing its path runtime and looking it up in an std::map.

Improves a data loading benchmark with ~3%, in addition to the ~7% regression
it fixes.

Change-Id: Ia536f6342278fcd6cc990053c3d2b0978e781b29