Bug#27205252: VALGRIND: UNINITIALIZED VALUE IN COPY_FIELD::GET_COPY_FUNC
When running the json_conversions test under Valgrind with release
binaries, a "Conditional jump or move depends on uninitialised
value(s)" warning is raised by Valgrind for the following statement:

create table t_geo select DISTINCT(_geo) FROM at;

This is a false positive caused by an optimization in GCC. Valgrind is
correct that uninitialized memory is read, but it is only done because
the compiler is reordering statements after it has been able to prove
that reading possibly uninitialized memory won't affect the result.
The Valgrind documentation says that such false positives can happen
with high optimization levels.

It is an optimization in Nullable<gis::srid_t>::operator==() that
triggers the warning. The code checks that both the Nullable objects
contain a value before it tries to compare their values, but the
compiler seems to lay out code that compares the values before the
code that checks if they contain a value.

This patch silences the Valgrind warning by using value-initialization
instead of default-initialization of Nullable::m_value when no value
has been provided.

Change-Id: Ibb039cfc45020ece4b295193d9d8c46c4da9103d