Bug#27149369: ASSERTION FAILURE IN BOOL               READ_FRAME_BUFFER_ROW(INT64, WINDOW*, BOOL)
Problem:
When handling range frames, if the first row for a range frame is found,
currently we do not store the position - REA_FIRST_IN_FRAME(as a hint).
This affects how we retrieve the row from frame buffer and fails the assert.

Analysis:
For these set of values - 392, 1027,1027,1034,1039 and window having
RANGE BETWEEN 7 PRECEDING and 1 PRECEDING,
For the fourth row, when calculating the rows that are in range, 1027 from
row 3 will be marked as first in frame and 1027 from row 4 will be
marked as last in frame. As a result when we get back to calculating
"last_value" for row 5 (value 1039)- Since we have first_row < last_row
in the frame for the previous range frame, we try and retrieve
FIRST_ROW_IN_FRAME from the frame buffer. But we fail to find the row using
the hints present in buffer_positions.

Solution:
Store the hint REA_FIRST_IN_FRAME when we find the first row in the range
frame as this will be used for the next range frame calculation.