BUG#27212924: GROUP_REPLICATION.GR_START_WITH_PERSISTED_VARIABLES TEST FAILS
When a member joins a group, all members, including the joiner, do
log a View_change_log_event transaction, than on the joiner the
generated locally is skipped by the GTID auto skip behaviour.
The joiner will apply the View_change_log_event transaction that is
delivered through recovery stage and the skip the local one.
If the server is stopped in the middle of these events, the local
generated View_change_log_event transaction will remain on the relay
log to be skipped later.
Though, on MTR clean between tests everything is clean except the
group_replication_applier channel, which due to the extra
transactions from a previous group was making the START
GROUP_REPLICATION to fail.

To avoid that situation, now we also clean group_replication_applier
channel between tests.