Bug#27220467 SERVER HANGS WHILE CREATING A NEW SRS WHICH IS USED IN ANOTHER CLIENT
When adding/modifying a column with the SRID property, the server will take
an MDL lock on the SRID in order to check if the spatial reference system
actually exists or not. In case of any error while adding/modifying the column,
the lock would not be released. This patch explicitly releases any locks taken
during adding/modification of columns.

Change-Id: I381b787f5912c9d1d7c55f5bad4178dca4936f23