bug#24444908  CLUSTER CRASHED DURING RESTART WITH AN ASSERTION `!CALLBACKOBJ->HAS_DATA_TO_SEND.
Addendum patch to fix a regression introduce by previous patch set:

Patch for this bug introduced enable/disable semantics for send_buffers. Initially
all send buffers are disabled, and got enabled as Transporters 'connect', or
when we 'open' a ned trp_client.

However, if the 'open' of a new client required the m_threads[] array to
be expanded, it have to send a request signal to its own clusterMgr.
This happened *before* we had set up which nodes send_bufffers were enabled
for this client. Thus the send (silently) failed, and 'open' waited
forever for the m_threads[] array to be expanded.

This patch reordered the 'open' logic, such that the TransporterFacade set
of already enabled send_buffers are assigned to the to-be-opened-client
before we possible send the expand request.