BUG#27179336 - UNINSTALL GROUP REPLICATION CAN DEADLOCK WITH GR START ON SERVER START
Issue:
------
If the server is booted with group replication automatically starting, the
control is returned to the user when the server is still in read mode.
At this point, if the user decides to unset the read only mode and uninstall
the plugin, the latter command deadlocks.

This happens due to two reasons:
1) The uninstall request gets stuck because it is waiting on GCS to terminate.
2) GCS gets stuck during the join phase, because it notices that the read
mode is off and tries to enable it. However, it isn't able to set it because
it is still waiting on the uninstall request.

Fix:
----
To avoid the deadlock during the uninstallation of Group Replication,
the value of the new plugin_is_setting_read_mode flag is verified in the
plugin_group_replication_check_uninstall function. If it is true, the
uninstallation fails, as the plugin is still busy setting the read
mode. Otherwise, the uninstallation proceeds normally.