WL#9059: [1/8] Dictionary table modifications
Create the data dictionary entry for SRID 0 when the data dictionary is
initialized instead of using an INSERT statement. This means that when
the dictionary table is hidden and INSERT is replaced by CREATE SPATIAL
REFERENCE SYSTEM, we don't have to support creating SRID 0.

Rename the UNIQUE index on name to SRS_NAME to get a better ER_DUP_ENTRY
error message when trying to create two SRSs with the same name.

Set the collation on organization to a case insensitive collation and
add a UNIQUE index on organization and organization_coordsys_id to
enforce a requirement in SQL/MM 2015, Part 3, Sect. 20.3. Remove
organization and organization_coordsys_id specifications from SRSs
created in tests to avoid duplicates.

Change-Id: I2642f0fa88f5e457c929649f0378a4891e6645ab