Bug #27214153: GCOL: INCORRECT BEHAVIOR WHILE INSERTING                      INTO A VIEW
Issue:
------
Before inserting into a table or view,
validate_gc_assignment checks if any generated columns
are being inserted into. Then it will raise the
appropriate error.

At this stage columns from a view are of the class
Item_direct_view_ref (instead of Item_field if it were a
table), so extracting the field object might cause issues.

Solution:
---------
Use the real_item() function to avoid this problem. It will
always give the underlying object being referred to by a
view.

This problem was fixed in trunk by Bug#22376872.