Bug #26864199: INCORRECT BEHAVIOR WITH VIEWS IN SP
Issue:
------
This problem occurs when an SP contains a statement that
refers to a VIEW, which in turn refers to another VIEW.

For views, a LEX object is created and retained in
"parent_lex" when handling the view definition which is a
query expression. While testing whether a
derived-table / view should be merged
(in SELECT_LEX::merge_derived), due to this LEX object it
incorrectly assumes that the current SELECT_LEX object is
the outer-most in the query and doesn't proceed further.

It would eventually result in not creating a field for the
view using create_view_field. This causes incorrect
behavior with second execution of the stored procedure.

Solution:
---------

On 5.7:
Fix is a backport of Bug #26627136 (pushed to 8.0.3 as part
of WL#6526).

A new function has also been added to restore the default
LEX object and remove any reference to the previous LEX
object created while parsing a view.

On trunk and 8.0:
Chaning the name of the function restore_parent_lex to
merge_query_blocks. And the parameter name from old_lex to
parent_lex.