BUG#26135313: MTR IS NOT COPYING .LOG FILE TO FAILURE DIRECTORY
Issue:
------
In MySQL-5.6, when a test case fails, MTR creates a failure specific
directory under 'var/log' and copies all the test specific contents
like data directory, .cnf file, 'tmp' directory etc to that location,
but it doesn't copy the log file created.

From MySQL-5.7, the log file created is also copied to failure
specific directory and then the correct location of it is printed on
the console.

Fix:
----
- Copy the log file created to the test failure specific directory and
  then print the path to it on the console.

- Removed the uneccessary code creating testname.log.failed file which
  is the copy of the log file, since the log file is now copied to the
  failure specific directory when there is a test failure.

Change-Id: I2b9c3964f693f3857c3e47ef62d7aebc559b9d03