Bug#27266086 	POLLEVENTS() 'MAX_WAIT' MAY OVERSLEEP UP TO 9MS
As the max wait time was waited for in chunks of max 10ms, we could
have less than 10ms remaining in the last wait-nap. However, we
still waited for 10ms, which could result in up to a 9ms
over sleep.

Patch recalulate the sleep  time for each nap, and wait less than
10ms if less than 10ms remains.