Bug#27130570 : ASSERT IN DBLQH::SENDKEYINFO20()
Dblqh::sendKeyinfo20() method skips amending the actual keyinfo, if the
request is from a SUMA block from the kernel as the block is only
interested in the scanInfo_Node detail. The sender BlockReference is
used to find out if the request is from a SUMA block. But in certain
circumstances where multiple NDB objects are being used parallely from
an API node, the block number extracted from the blockreference is same
as that of a SUMA block even though the request is coming from an API
node. Due to this ambiguity, the Dblqh mistakes the request from the
API node for a request from a SUMA block and crashes. This patch fixes
that by checking node ids before checking the block number.

(Already pushed to higher versions so null merging up this commit)