Bug #27218248 	MOVE THE 'DEFAULT' SEND_BUFFER HANDLING OUT OF 'CLASS TRANSPORTERREGISTRY'
Pure refactoring patch:

'class TransporterRegistry' is a common component used both internally in the
data nodes, and as part of the NDB API transporter stack. It 'collects' the
different aspects of possibly multiple Transporter implementations, and
different send- and receiver-buffer providers and consumers used by either
multi-threaded or single-threaded ndb, or the NDB API.

Even though this is intended to be a general components, it also has a
'default' implementation of the send-buffers. This is only used by the
single-threaded ndb, so this code is really misplaced in class
TransporterRegistry.

This patch moves the 'default' send_buffer implementation out of
class TransporterRegistry and into class TransporterCallbackKernelNonMT.
This class already contain other non-mt kernel implementations and seems
like a natural harbour for this stuff.