Bug #25949639: DATE FORMAT 'YYYYMMDD' ISN'T RECOGNIZED IN                LEFT JOIN
Issue:
------
This problem occurs when DATE (or related) column / value
is compared to a string in the format 'YYYYMMDD'. When a
DATETIME column is involved, values shouldn't be compared
as strings since it won't take the format into account.

Similar problems were found in CASE and IN statements.

Solution:
---------
If DATE (or related) values are involved, use
"cmp_item_datetime" for comparision. This will take the
relevant format into account.

To handle TIME values "cmp_item_datetime" has a new flag
to call "get_time_value" for that case. Also making
warn_item into a const to make sure the object isn't
modified.