WL#9185 MySQL Cluster support for new DD
Change ndb_reconnect to require that binary logging is off since
that's what test is designed for. Remove warning suppressions which
would occur only if binary logging was on, this reduces clutter in the
test. Remove check for not windows, there is nothing in the test that
does not work on Windows nowadays. Improve comments.

Change-Id: I1580771ea47e7d40cf6117dd856699aa9f0e85b5