WL#8648 NDB_SHARE lifecycle improvements
Remove redundant error logging when failure to create event, event
data or event operation occurs in the code path that setup
event subscription on a table. The error is redundant since
a warning with same message has already been logged and the caller
checks the return code of the function call, thus the caller
controls which action to take upon failure.

Change-Id: Ibac3afd1430e3d0fbb59d86c9b15ad92c7bd4398