Bug#27096812 - MISLEADING ERROR MESSAGE WHEN TRYING TO REMOVE A NODE THAT DOES NOT EXIST
Description
-----------
If GCS tries to remove a node after getting a notification that a node is
faulty but the node is removed before it can act upon the information, a
misleading message is printed out:

2017-11-05T14:33:47.142598Z 0 [Note] Plugin group_replication reported:
'Old incarnation found while trying to remove node localhost:13090
15098924083888470.'

This may happen, for example, while forcing a new configuration after a
majority loss.

Solution
--------
Changed the messaged printed when the node has already been removed to:
"Node has already been removed: <node address> <node UID>."
In addition to changing the message printed, we now also prevent the
'remove_node' request from proceeding when the node has already been
removed.
In addition, added a message informing the user when the node to remove
has reincarnated with a new UID; as before, we do not allow an upper layer
to remove a reincarnated node.