Bug#25701141 - INSERT WITH CURSOR_TYPE_READ_ONLY HANGS
DESCRIPTION:
============
When trying to execute an INSERT prepared statement with
CURSOR_TYPE_READ_ONLY set, libmysqlclient will hang forever.

ANALYSIS:
=========
When INSERT prepared statement is executed and
COM_STMT_EXECUTE is sent to the server, the server just
sends an OK packet in this case. This packet is read by the
client while reading the query result. However, after this,
since CURSOR_TYPE_READ_ONLY is set, the client attempts to
read additional packet although there is no packet left to
be read. This causes libmysqlclient to hang.

When the CURSOR_TYPE_READ_ONLY is set, we can expect an
additional data or OK packet mainly for Prepared
statements having SELECT queries or when we are expecting
a result set as a part of query output. The SELECT query
output could be in the form of <metadata><OK> or
<metadata><rows><OK>. Since the metadata packet will be read
while reading the query result and mysql->field_count would
be populated by then, we can decide based on that whether
to expect more packets or not.

FIX:
====
The fix is to make sure that we perform appropriate checks
before performing additional packet reads in cases where a
read-only cursor is set. This would make sure that the
client does not hang waiting for a packet from the server.