WL#8069 Post-push fix BUG#27245083 We are using two different methods to distinguish whether a page have already been handled by LCP scans. The first one is a flag on the page. This flag is set when the first alloc/dealloc of a page during an LCP scan is an allocation, in this case set the page flag to indicate that the page can be skipped since we had no rows in the page at the start of the LCP scan. The second variant is used when the first event is a release of the page. In this case we set the lcp_scanned_bit in the page map. Later alloc/release of the same page will not touch any of those flags.
The reason that we have two approaches dependent on whether the
first event is alloc or release is that we need to know the page
state at the start of the LCP scan to properly set the last LCP
state in each page during the LCP scan. If we only use one flag
we will lose the vital information about what the state of the
page at the start of LCP was.

(cherry picked from commit a5c75db7af7916d2a91b324ad4079ee32e4e53fd)