Bug#27015112: RPL.RPL_MTS_LOGICAL_CLOCK_APPLIER_ERROR_CONCURRENT_STOP_SLAVE UNSTABLE ON PB2
  The test was failing with Timeout in include/wait_for_slave_param.inc.
The test syncs the IO thread and then checks that the Slave_SQL_Running_State
is 'waiting for dependent transaction to commit'. But before the test
checks it on slow platforms the lock_wait will timeout as a result SQL_THREAD
will be stopped hence status will not match.

Fix. Stopped the SQL_THREAD before doing the transaction which will cause
SQL status on slave as 'waiting for dependent transaction to commit'.Started
the SQL_THREAD just before checking for Slave_SQL_Running_State.