Bug #25076416   VIRTUAL COLUMN IS NOT CONSIDERED WHILE                 FETCHING THE AUTOINC COLUMN
Analysis:

While fetching the auto-increment value from the table, InnoDB
tries to find the index which has the first column as autoincrement column.
While searching the index, InnoDB uses column offset of the table
to fetch the auto_increment column but virtual column is not considered
while searching the column and it leads to wrong column.

Fix:

Consider the virtual column of the table while fetching the
auto_increment column from the table using column offset.

Reviewed-by: Jimmy Yang<Jimmy.yang@oracle.com>
RB: 16281