Bug#26040870 - ASSERT ON KILL'ING A STORED ROUTINE INVOCATION
Issue here is, execution of the stored routine statement's
is interrupted by the KILL operation. In this case diagnostics
area is not set but it is accessed to check the error reported by
mistake. Hence assert condition to check if error is reported in
the diagnostics area is hit.

Patch for the Bug25586773, fixes this issue in all the reported
versions by adding check to verify if diagnostics area is set
before accessing it.

As part of this patch, only test case is added for the coverage.

Thanks to Laurynas Biveinis for reporting a bug and contributing
patch for it.

Change-Id: I52d724154be5bfab2f0769915142fb441d61c174