BUG #25548088: --SSL-MODE SUPPORT IN MEB SHOULD BE THE SAME AS IN MYSQL CLIENT
Description:

Bug#23508228 - --SSL-MODE NEED TO BE SUPPORTED BY MEB, introduces --ssl-mode to
MEB in 4.1.0 version. It's implementation restricts the values to this option
to be only in uppercase. This is not the same behavior for mysql client. mysql
client accepts the values which are case-insensitive. So currently, MEB
mandates the uppercase for the value of --ssl-mode.

Example:
--ssl-mode=REQUIRED --> "Accepted"
--ssl-mode=required --> "Rejected"

Fix:

Use 'find_type_or_exit' method to determine the type of ssl-mode requested by
the user. MySQL client uses the same function call to determine the type of
ssl-mode requested by the user.

Existing method 'get_ssl_mode_by_name' can be discarded as it can no longer be
used by MEB.

'find_type_or_exit' terminates once an invalid value is specified, with a
proper message explaining the possible values for a given option. Currently,in
MEB an error message is thrown if an invalid value is specified for ssl-mode.
This error message can no longer be observed with this implementation.