BUG#27165446: PARALLEL REPLICATION OF CREATE/ALTER/DROP DATABASE DEADLOCK, EARLY MDL RELEASE
DDL statements may take a shared SCHEMA metadata lock in case the table
definition is not in the table definition cache. This lock is not taken
in a deterministic manner: it may happen that a replication slave takes the
lock even if the master did not take it. When a replica uses
slave-presere-commit-order=1, we suspect that this can lead to a deadlock, as
follows:

- The master executes and commits ALTER TABLE and ALTER DATABASE
concurrently, with no lock conflict.
- The binlog order becomes:
    ALTER TABLE
    ALTER DATABASE
- The master tags the two transactions as non-conflicting so that the
replica can apply them together.
- On the slave, ALTER DATABASE executes first.
  - Take exclusive lock on SCHEMA MDL
  - Execute
  - Due to slave-preserve-commit-order wait for ALTER TABLE to commit
- Then ALTER TABLE executes
  - Finds table is not in the table definition cache
  - Transaction needs the SCHEMA MDL.

Since ALTER DATABASE already holds an exclusive lock on the SCHEMA MDL while
waiting for ALTER TABLE to commit and ALTER TABLE needs SCHEMA MDL there is
a deadlock.

Resolution:
Make replication force CREATE/ALTER/DROP DATABASE to execute in isolation.