Bug#27343113 MTR: --SKIP-RPL WONT SKIP FEW TESTS IN RPL_GTID SUITE
Issue:
------
Using the MTR option --skip-rpl skips tests which are in rpl suite
even though it does not include master-slave.inc/rpl_init.inc. But,
--skip-rpl fails to skip tests which are in rpl_gtid suite and does
not include master-slave.inc/rpl_init.inc.

This is a regression caused by the patch for Bug#25166686 due to
which only tests in the rpl suite were being tagged as rpl_test's
but not those in rpl_gtid or rpl_nogtid suites.

Fix:
----
Identify all suites whose names match 'rpl' as replication tests
and skip them when the option --skip-rpl is specified.

Change-Id: I505cef82a7dfc0d5528dc588068f4347d2fc3e61