Bug#27348276 WRONG RESULT FOR WINDOW FUNCTION FIRST_VALUE WITH EMPTY FRAME
The value of Item_first_last_value::maybe_null was set from the
corresponding value of its argument. In this case, it is false.

If we have an empty frame the value of FIRST_VALUE and
LAST_VALUE should be NULL. The correspondig null_value was set by the
logic for this wf.  Item_first_value_value::compute returns true in
two cases: a) an error has occured, so return a default value, or b)
we have a null value so no further processing is required, i.e. we can
return early.  The null_value, however, was changed by our use of
error_int, error_str in this case, which inspects maybe_null to
determine what default value to return.

Fix: make sure maybe_null is always true for Item_first_last_value
notwithstanding its argument.