Bug #27349285 	CLUSTERMGR 'IS_CONNECTED' STATE INCORRECTLY REPORTED
The ClusterMgr is responsible for *synchronously* keeping track of the connection state for all nodes.
In order to achive that, the methods ClusterMgr::reportConnected() and ::reportDisconnected()
is called when ::update_connections() detect a changed connection state
*with the ClusterMgr::m_mutex* locked.

That should cause the ClusterMgr::theNodes[] states to be updated accordingly,
which could later be checked by methods as ::is_connected() to determine the state.

It is reasonable to expect that immediately after:
1) ... calling ::reporteConnected(), ::is_connected() will return true.
2) ... calling ::reportDisconected(), ::is_connected() will return false.

However, it turns out that 2) does not behave as expected. This is due to
::reportDisconnect() not directly updating ::theNodes[] itself. Instead it send
a DISCONNECT_REP signal to itself, and leaves it to ::execDISCONNECT_REP() to
set the disconnected state later. Thus, until this signal is executed, the
is_connected() state will be incorrectly reported.

Patch moves the update of theNodes[] state from execDISCONNECT_REP into
::reportDisconnect(), this is similar to how ::reportConnect() is implemented.
Note, that it should be safe to do this update directly in ::reportConnect/Disconnect
as they are called by the poll owner, protected by holding the m_mutex() (assert added)

::execDISCONNEC_REP becomes obsolete and is removed.