Bug#26399073 - MYSQL DOES NOT COMPILE WITH CLANG ON WINDOWS
Fix a performance regression; InnoDB has a system where the performance
schema key is decided by FILE, and the patch that extended the basename
function to work with both / and \ made these allocations slower (as it
uses std::string, which can allocate on the heap).

However, there is no reason whatsoever that this lookup needs to be done
runtime in a map. Change it to use a constexpr function; it goes through
the array linearly, but it's small and it's compile-time. The end result
is that the compiled code stores a PSI key directly, instead of taking
FILE, removing its path runtime and looking it up in an std::map.

Improves a data loading benchmark with ~3%, in addition to the ~7%
regression it fixes

rb#18379 Approved by Bin. Fixed by Steinar.