Bug #27360492 DEADLOCK DETECTED AFTER WL#8960
Problem:

The problematic code is in lob::purge():

338         first_page_t first(mtr, index);
339         first.load_x(page_id, page_size);
340
341         page_type_t page_type = first.get_page_type();
342
343         if (page_type == FIL_PAGE_TYPE_ZBLOB
344             || page_type == FIL_PAGE_TYPE_BLOB
345             || page_type == FIL_PAGE_SDI_BLOB
346             || page_type == FIL_PAGE_SDI_ZBLOB) {
347                 lob::Deleter free_blob(*ctx);
348                 free_blob.destroy();
349                 DBUG_VOID_RETURN;
350         }

In line 339, a latch is taken on the first page of LOB.  And when we
enter the if condition in 343, we continue to hold this latch.
The Deleter::destroy() will free the full LOB using a local mtr.  So
the latch taken in 339 is still in place.  When another thread tries to
allocate that page, a deadlock is reported.  I think this is the
root cause of the problem.

Solution:

When Deleter::destroy() is called, ensure that no mtr is holding latch
on the first page of LOB.