Bug #27264789: CONCURRENT EXECUTION OF SET/RESET PERSIST CALL CRASHES SERVER
Problem: When SET PERSIST/RESET PERSIST are run in parallel there is a crash.
This is because the data structure(m_persist_variables/m_persist_ro_variables)
used to hold all persisted variables is not protected with a mutex in case of
concurrent update.

Fix: Fix is to protect m_persist_variables/m_persist_ro_variables variables
with a mutex.